+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- abort ingestion of further items after client disconnects
+
 ### Deprecated
 
 ### Removed