+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - abort ingestion of further items after client disconnects
 - wrap neo4j errors during ingestion in ingestion errors
+- bumped cookiecutter template ed5deb
 
 ### Deprecated